feat(FIN-030): removed lazy dropdown
