+++ package.json
@@ -17,6 +17,7 @@
   "author": "Oleksandriia's Frontend Team",
   "license": "ISC",
   "dependencies": {
+    "@base-ui/react": "^1.3.0",
     "@formkit/auto-animate": "^0.9.0",
     "@hookform/resolvers": "^5.2.1",
     "@hugeicons/core-free-icons": "^4.0.0",
