+++ src/client/shared/components/fields/props/input.props.ts
@@ -1,7 +1,6 @@
 import { Dispatch, SetStateAction } from 'react';
 import { InputDefaultProps } from '../../../props/input-default.props';
 import { DropdownDefaultProps } from '../../../props/dropdown-default.props';
-import { LazyEventProps } from '../../../props/lazy-event.props';
 
 export interface InputProps<T> extends InputDefaultProps {
   onChange: Dispatch<SetStateAction<T | undefined>>;
@@ -9,13 +8,6 @@ export interface InputProps<T> extends InputDefaultProps {
 
 export type DefaultDropdownInputProps<T> = Omit<InputProps<T>, 'value'> &
   DropdownDefaultProps<T> & {
-    value?: T | undefined | null;
-    customInputValue?: string | undefined | null;
-  };
-
-export type LazyDropdownInputProps<T> = Omit<InputProps<T>, 'value'> &
-  DropdownDefaultProps<T> &
-  LazyEventProps & {
     value?: T | undefined | null;
     customInputValue?: string | undefined | null;
   };
